requirements.txt
pyserial
tkeasygui